Pass requested CUDA_VERSION and LINUX_VERSION to dind build
If the user specifies a LINUX_VERSION and/or CUDA_VERSION in their
environment, we need to use it when building the dind container.
Otherwise we can get in a situation where the host driver is older
than the defaulted 11.8.

I don't really know what I'm doing here, but this allowed me to build
a container on where the host driver is 11.6 and I had requested that
in the environment.