Don't delete RMM's `_torch_allocator.cpp`
The RMM Python source now contains a non-generated C++ file,
`_torch_allocator.cpp`, from rapidsai/rmm#1407, so we need to avoid
deleting it when cleaning rmm-python. This approach might not be the
best if RMM Python goes on to include more C++ sources, but this fixes
the clean / build process for now.